+++ library/src/main/java/org/mustangproject/Item.java
@@ -1,5 +1,6 @@
 package org.mustangproject;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import org.mustangproject.ZUGFeRD.IReferencedDocument;
@@ -301,12 +302,14 @@ public class Item implements IZUGFeRDExportableItem {
 		return this;
 	}
 
-	@Override public IZUGFeRDAllowanceCharge[] getAllowances() {
+	@JsonIgnore
+	@Override public IZUGFeRDAllowanceCharge[] getAllowances() { // in JSON is already returned as itemAllowances (and only read from there)
 		IZUGFeRDAllowanceCharge[] izac=new IZUGFeRDAllowanceCharge[Allowances.size()];
 		return Allowances.toArray(izac);
 	}
 
-	@Override public IZUGFeRDAllowanceCharge[] getCharges() {
+	@JsonIgnore
+	@Override public IZUGFeRDAllowanceCharge[] getCharges() { // in JSON is already returned as itemAllowances (and only read from there)
 		IZUGFeRDAllowanceCharge[] izac=new IZUGFeRDAllowanceCharge[Charges.size()];
 		return Charges.toArray(izac);
 	}