+++ History.md
@@ -13,7 +13,9 @@
 - #874
 - #876/830
 - #878
-- corrected typo ArithmetricException to ArithmeticException 
+- corrected typo ArithmetricException to ArithmeticException
+- 726
+- #885
 
 2.17.0
 =======